Add Components for basic calculation language
Added Components & Visitors:
   NumberComponent
   AdditionComponent
   SubtractionComponent
   MultiplicationComponent
   DivisionComponent (Need to resolve divided by 0 issue)
   PowerComponent
   NegativeComponent (Need to pass down the value, but maybe hard to implement)
   ParenthesesComponent
   BlockComponent

Current Issues:
   DivisionComponent: Need to resolve divided by 0 issue
   NegativeComponent: Unable to pass the negative down

+++ src/main/java/flitesharp/component/literal/LiteralComponent.java
@@ -21,7 +21,7 @@ public abstract class LiteralComponent extends Component {
      * @return the numeric value of a literal if any
      * @throws UnsupportedOperationException if the literal hasn't a numeric value
      */
-    public float getNumberValue() throws UnsupportedOperationException {
+    public double getNumberValue() throws UnsupportedOperationException {
         throw new UnsupportedOperationException();
     }
 }
